+++ CHANGES.md
@@ -11,6 +11,7 @@ CHANGELOG
 * Parallel ways drawing
 * A4 and A3 formats
 * Add address search tool
+* Draw an optional arrow at the end of a line
 
 **Minor changes**